Repair constructors and some other fixes
The constructors where broken due to a misunderstanding of their syntax
(especially with parametrised types).

Applying a kernel to an array now work for Lanczos kernels.

`iscardinal` and `isnormalized` methods work on instances and on some types
(when these properties do no depend on the value of some kernel coefficients).

`convert` works and is more consistent.